requirements.txt
ultralytics
opencv-python
python-telegram-bot
